+++ logger.py
@@ -1,0 +1,7 @@
+import logging
+
+# 設定 logging
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+)
+logger = logging.getLogger(__name__)
